requirements.txt
accelerate==0.26.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
av==14.3.0
braceexpand==0.1.7
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
datasets==3.5.0
decorator==5.2.1
diffusers==0.33.1
dill==0.3.8
docker-pycreds==0.4.0
einops==0.6.1
einops-exts==0.0.4
executing==2.2.0
EzColorLog==1.0.3
fastapi==0.115.12
filelock==3.13.1
frozenlist==1.5.0
fsspec
ftfy==6.3.1
gcsfs==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos
h11==0.14.0
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jaxtyping==0.3.1
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
latex2mathml==3.77.0
markdown2==2.5.3
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
networkx==3.3
numpy==1.26.3
oauthlib==3.2.2
omegaconf==2.3.0
open_clip_torch==2.32.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
peft==0.14.0
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.7
prompt_toolkit==3.0.50
propcache==0.3.1
proto-plus==1.26.1
protobuf
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
qwen-vl-utils==0.0.10
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.2.2
scipy==1.15.2
sentencepiece==0.1.99
sentry-sdk==2.25.1
setproctitle==1.3.5
shortuuid==1.0.13
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
starlette==0.46.1
svgwrite==1.4.3
sympy==1.13.1
threadpoolctl==3.6.0
timm==0.9.16
tokenizers==0.21.0
torch==2.5.1
torchaudio
torchvision
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
typeguard==4.4.2
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.0
wadler_lindig==0.1.4
wandb==0.19.9
wavedrom==2.0.3.post3
wcwidth==0.2.13
webdataset==0.2.86
wget==3.2
xxhash==3.5.0
yarl==1.19.0
zipp==3.21.0
